#4 [test] Product 컨트롤러 테스트 수정 - jsonPath -> response 객체로 받아서 검증
